Полноценная сцена с поддержкой физики Bullet #107
