+++ CMakeLists.txt
@@ -132,6 +132,11 @@ if(NOT IL_FOUND)
     message(FATAL_ERROR "DevIL library not found..")
 endif()
 
+find_package(Bullet)
+if(NOT BULLET_FOUND)
+    message(WARNING "Bullet library not found.. Physics simulation sample will be disabled")
+endif()
+
 include_directories("${PROJECT_SOURCE_DIR}/lite3d/include")
 include_directories("${PROJECT_SOURCE_DIR}/lite3dpp/include")
 include_directories("${PROJECT_SOURCE_DIR}/lite3dpp_font/include")
@@ -139,6 +144,11 @@ include_directories("${PROJECT_SOURCE_DIR}/lite3dpp_font/include")
 add_subdirectory("${PROJECT_SOURCE_DIR}/lite3d/")
 add_subdirectory("${PROJECT_SOURCE_DIR}/lite3dpp/")
 add_subdirectory("${PROJECT_SOURCE_DIR}/lite3dpp_font/")
+
+if(BULLET_FOUND)
+    include_directories("${PROJECT_SOURCE_DIR}/lite3dpp_physics/include")
+    add_subdirectory("${PROJECT_SOURCE_DIR}/lite3dpp_physics/")
+endif()
 
 if(BUILD_TESTS)
     add_subdirectory("${PROJECT_SOURCE_DIR}/tests/")
