+++ lite3dpp_physics/include/lite3dpp_physics/lite3dpp_physics_motion_state.h
@@ -15,29 +15,25 @@
  *	You should have received a copy of the GNU General Public License
  *	along with Lite3D.  If not, see <http://www.gnu.org/licenses/>.
  *******************************************************************************/
-#include <SDL_assert.h>
+#pragma once
 
-#include "lite3dpp_physics_motion_state.h"
+#include <lite3dpp_physics/lite3dpp_physics_bullet.h>
 
 namespace lite3dpp {
 namespace lite3dpp_phisics {
 
-PhysicsObjectMotionState::PhysicsObjectMotionState(SceneObject *o) : 
-    mSceneObject(o)
-{}
+    class PhysicsObjectMotionState : public btMotionState
+    {
+    public:
 
-void PhysicsObjectMotionState::getWorldTransform(btTransform& worldTrans) const
-{
-    SDL_assert(mSceneObject);
-    worldTrans.setOrigin(BulletUtils::convert(mSceneObject->getPosition()));
-    worldTrans.setRotation(BulletUtils::convert(mSceneObject->getRotation()));
-}
+        PhysicsObjectMotionState(SceneObject *o);
 
-void PhysicsObjectMotionState::setWorldTransform(const btTransform& worldTrans)
-{
-    SDL_assert(mSceneObject);
-    mSceneObject->setPosition(BulletUtils::convert(worldTrans.getOrigin()));
-    mSceneObject->setRotation(BulletUtils::convert(worldTrans.getRotation()));
-}
+        void getWorldTransform(btTransform& worldTrans) const override;
+        //Bullet only calls the update of worldtransform for active objects
+        void setWorldTransform(const btTransform& worldTrans) override;
 
+    private:
+
+        SceneObject *mSceneObject;
+    };
 }}
