+++ lite3dpp_physics/src/lite3dpp_physics_scene_node.cpp
@@ -15,7 +15,7 @@
  *	You should have received a copy of the GNU General Public License
  *	along with Lite3D.  If not, see <http://www.gnu.org/licenses/>.
  *******************************************************************************/
-#include "lite3dpp_physics_scene_node.h"
+#include <lite3dpp_physics/lite3dpp_physics_scene_node.h>
 
 namespace lite3dpp {
 namespace lite3dpp_phisics {
