+++ samples/CMakeLists.txt
@@ -32,13 +32,6 @@ define_sample(sample_vault_111 sample_common lite3dpp_font lite3dpp)
 define_sample(sample_vault_room sample_common lite3dpp_font lite3dpp)
 define_sample(sample_reflection_sky sample_common lite3dpp_font lite3dpp)
 
-find_package(Bullet)
-if(NOT BULLET_FOUND)
-    message(WARNING "Bullet library not found.. Physics simulation sample will be disabled")
-else()
-    set(SAMPLE_ADDON_INCLUDES ${BULLET_INCLUDE_DIRS})
-    set(SAMPLE_ADDON_LIBRARIES ${BULLET_LIBRARIES})
-    define_sample(sample_physics sample_common lite3dpp_font lite3dpp)
-    unset(SAMPLE_ADDON_INCLUDES)
-    unset(SAMPLE_ADDON_LIBRARIES)
+if(BULLET_FOUND)
+    define_sample(sample_physics sample_common lite3dpp_font lite3dpp lite3dpp_physics)
 endif()
